requirements.txt
# API y servidor
fastapi==0.135.1
uvicorn==0.42.0

# WebSocket — requerido por Uvicorn para manejar conexiones WS
websockets==14.1

# Base de datos MariaDB
aiomysql==0.2.0

# Redis
redis==5.2.1

# Ollama - cliente oficial Python
ollama==0.6.1

# Variables de entorno
python-dotenv==1.0.1

# Validación de datos
pydantic==2.12.5

# RabbitMQ — cliente asíncrono compatible con asyncio y FastAPI
aio-pika==9.4.3